backend/requirements.txt

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-filter==23.4
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0

# Advanced Features
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-ratelimit==4.1.0
django-cacheops==6.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
google-generative-ai==0.4.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Security
cryptography==41.0.4
pyotp==2.9.0
qrcode==7.4.2

# Monitoring
sentry-sdk==1.40.0
prometheus-client==0.17.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
bleach==6.0.0
lxml==4.9.3
